applied container/presentational to ItemArea
